pyproject.toml
[project]
name = "risenow-dspy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "dspy[mcp]>=3.0.3",
    "langfuse>=3.3.4",
    "openinference-instrumentation-dspy>=0.1.29",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.12",
    "textual>=6.1.0",
]
